refactor: rename notification module
